requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
itsdangerous==2.1.2
jsonschema==4.20.0
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
psycopg2==2.9.9
six==1.16.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
